1.3.1.1
- Code improvement

+++ Assets/Scripts/Shop/UpgradeSystem.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 public class UpgradeSystem : MonoBehaviour
 {
     public void UpgradeBase(UpgradeButton upgradeButton)
-    => BaseController.instance.GetComponent<EntityHealth>().MaxHealth = upgradeButton.UpgradeValue;
+    => BaseController.instance.GetComponent<EntityHealth>().MaxHealth = (int)upgradeButton.UpgradeValue;
 
     public void UpgradeGun(UpgradeButton upgradeButton)
     => BaseController.instance.character.gun.damageMultiplier = upgradeButton.UpgradeValue;
 
     public void UpgradeAmmos(UpgradeButton upgradeButton)
-    => BaseController.instance.GetComponent<EntityHealth>().MaxHealth = upgradeButton.UpgradeValue;
+    => BaseController.instance.GetComponent<EntityHealth>().MaxHealth = (int)upgradeButton.UpgradeValue;
 }
